# Implementation/JavaApplication6/src/main/java/com/bms/api/dto/UpdateAccountStatusRequest.java
package com.bms.api.dto;

import jakarta.validation.constraints.NotBlank;

public record UpdateAccountStatusRequest(
        @NotBlank String status,
        String reason,
        Integer actorId,
        String role) {
}